Auto stash before merge of "main" and "origin/main"

+++ Triangle/Assets/Scripts/ElementHandler.cs
@@ -33,8 +33,8 @@ public static class ElementHandler
         }
     }
 
-    public static double DamageConverter(Element ownElem, Element damagingElem)
+    public static float DamageConverter(Element ownElem, Element damagingElem)
     {
-        return 1;
+        return 1f;
     }
 }
